feat(suite-native): report skips of onboarding analytics
